src/lambda/requirements.txt
python-binance 
python-dotenv==1.0.1
Requests==2.32.3
